+++ NEWS.md
@@ -1,7 +1,7 @@
 # gsm.digitpref (development version)
 
 - Added `stackedDigitChart()` for visualizing digit distribution by group. (@swanggilead, #5)
-- Added `splitByPlaces()` to split a number into digits. (@jonthegeek, #6)
+- Added `splitByPlace()` to split a number into digits. (@jonthegeek, #6)
 
 # gsm.digitpref 0.0.0